Add Context Message Markers - Passing `true` for supplyContextMessage, will add a context message of `Serilog-Log4NetSink` scoped just for the log call under the `NDC` stack, which you can utilise for Log4Net filters and other purposes. See Log4Net documentation: https://logging.apache.org/log4net/release/manual/contexts.html#stacks
This can let you specify a filter on an appender in Log4Net. An example, disable the Seq appender for Log4Net if the context message is present. Useful to stop doubling up logs, for when you have Log4Net and Serilog both pushing to the same sink during migration of logging practices.